+++ Entities/Card.cs
@@ -5,8 +5,8 @@ namespace HeroServer
 {
     public class Card
     {
-        public int Id { get; set; }
-        public int AppUserId { get; set; }
+        public long Id { get; set; }
+        public long AppUserId { get; set; }
         public String CSToken { get; set; }
         public int TypeId { get; set; }
         public String Number { get; set; }
@@ -22,7 +22,7 @@ namespace HeroServer
         {
         }
 
-        public Card(int id, int appUserId, String csToken, int typeId, String number, int digits, DateTime expirationDate, String holder,
+        public Card(long id, long appUserId, String csToken, int typeId, String number, int digits, DateTime expirationDate, String holder,
                     DateTime createDateTime, DateTime updateDateTime, int status)
         {
             Id = id;
@@ -38,7 +38,7 @@ namespace HeroServer
             Status = status;
         }
 
-        public Card(int id, int appUserId, String csToken, int typeId, String number, int digits, DateTime expirationDate, String holder, int status)
+        public Card(long id, long appUserId, String csToken, int typeId, String number, int digits, DateTime expirationDate, String holder, int status)
         {
             Id = id;
             AppUserId = appUserId;
