+++ Entities/CatalogRecord.cs
@@ -4,11 +4,11 @@ namespace HeroServer
 {
     public class CatalogRecord
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public String[] Fields { get; set; }
         public String[] Filters { get; set; }
 
-        public CatalogRecord(int id, String[] fields, String[] filters)
+        public CatalogRecord(long id, String[] fields, String[] filters)
         {
             Id = id;
             Fields = fields;
